feat: add `delay`, `platform` and `realtime` info to departures (#16)
The MVG provides some additional information for each departure. While
`realtime` seems to be always provided, `delay` is only available if
`realtime` is `True` and might otherwise be omitted.
Additionally, `platform` is present only for some transport types. In my
testing this is `U-Bahn` and `S-Bahn`, but not for `Bus` and `Tram`.

---------

Co-authored-by: Martin Dziura <m.dziura@tum.de>